create form component Fixes #17 refactor:code quality filestructure
