+++ package.json
@@ -16,7 +16,7 @@
     "watch:css": "postcss src/tailwind.css -o src/index.css --watch",
     "env:dev": "cross-env NODE_ENV=development",
     "env:prod": "cross-env NODE_ENV=production",
-    "react-scripts:start": "react-scripts start",
+    "react-scripts:start": "sleep 5 && react-scripts start",
     "react-scripts:build": "react-scripts build",
     "start": "run-p env:dev watch:css react-scripts:start",
     "build": "run-s env:prod build:css react-scripts:build",
