+++ src/App.js
@@ -1,8 +1,10 @@
 import React from 'react';
+import LoginForm from './components/LoginForm';
 
 function App() {
   return (
     <div>
+      <LoginForm />
       <h1 className="text-2xl bg-lightblue">The Movie Bubble</h1>
     </div>
   );
